refactor(currency): merge format_number types into one file; consolidate all tests into currency_test.go
