+++ packages/i18nify-go/modules/currency/currency_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestUnmarshalCurrency(t *testing.T) {
@@ -21,7 +22,7 @@ func TestUnmarshalCurrency(t *testing.T) {
 }
 
 func TestMarshalCurrency(t *testing.T) {
-	expectedJSON := `{"currency_information": {"USD": {"name": "US Dollar", "numeric_code": "840", "minor_unit": "2", "symbol": "$", "physical_currency_denominations": ["1", "5", "10", "25", "50", "100"], "symbol_position": ""}}}`
+	expectedJSON := `{"currency_information": {"USD": {"name": "US Dollar", "numeric_code": "840", "minor_unit": "2", "symbol": "$", "physical_currency_denominations": ["1", "5", "10", "25", "50", "100"]}}}`
 
 	inputData := map[string]CurrencyInformation{
 		"USD": {
@@ -175,4 +176,210 @@ func TestFormatCurrency(t *testing.T) {
 			assert.Equal(t, tt.Eexpected, result)
 		})
 	}
+}
+
+// boolPtr is a helper to get a pointer to a bool literal.
+func boolPtr(b bool) *bool { return &b }
+
+func TestFormatNumber(t *testing.T) {
+	tests := []struct {
+		name     string
+		amount   interface{}
+		opts     NumberFormatOptions
+		expected string
+		wantErr  bool
+	}{
+		{
+			name:     "USD prefix symbol with grouping",
+			amount:   1000.5,
+			opts:     NumberFormatOptions{Currency: "USD"},
+			expected: "$1,000.50",
+		},
+		{
+			name:     "negative USD",
+			amount:   -500.0,
+			opts:     NumberFormatOptions{Currency: "USD"},
+			expected: "-$500.00",
+		},
+		{
+			name:     "zero USD",
+			amount:   0,
+			opts:     NumberFormatOptions{Currency: "USD"},
+			expected: "$0.00",
+		},
+		{
+			name:     "plain decimal no currency",
+			amount:   "750.75",
+			opts:     NumberFormatOptions{},
+			expected: "750.75",
+		},
+		{
+			name:     "large number plain decimal",
+			amount:   12345.6789,
+			opts:     NumberFormatOptions{},
+			expected: "12,345.679",
+		},
+		{
+			name:     "JPY zero fraction digits",
+			amount:   5000,
+			opts:     NumberFormatOptions{Currency: "JPY"},
+			expected: "¥5,000",
+		},
+		{
+			name:   "disable grouping",
+			amount: 1234567.89,
+			opts: NumberFormatOptions{
+				UseGrouping: boolPtr(false),
+			},
+			expected: "1234567.89",
+		},
+		{
+			name:     "EUR prefix symbol",
+			amount:   1234.56,
+			opts:     NumberFormatOptions{Currency: "EUR"},
+			expected: "€1,234.56",
+		},
+		{
+			name:    "invalid string amount",
+			amount:  "not-a-number",
+			opts:    NumberFormatOptions{Currency: "USD"},
+			wantErr: true,
+		},
+		{
+			name:    "string with commas (not parseable)",
+			amount:  "1,234",
+			opts:    NumberFormatOptions{},
+			wantErr: true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := cachedCurrencyData.FormatNumber(tt.amount, tt.opts)
+			if tt.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, tt.expected, got)
+		})
+	}
+}
+
+func TestFormatNumber_IntlMappedCurrencies(t *testing.T) {
+	cases := []struct {
+		currency string
+	}{
+		{"SGD"},
+		{"AUD"},
+		{"HKD"},
+		{"CAD"},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.currency, func(t *testing.T) {
+			result, err := cachedCurrencyData.FormatNumber(1000.0, NumberFormatOptions{
+				Currency: tc.currency,
+			})
+			require.NoError(t, err)
+
+			sym, symErr := GetCurrencySymbol(tc.currency)
+			require.NoError(t, symErr)
+			assert.Contains(t, result, sym,
+				"formatted string should contain canonical symbol %q for %s", sym, tc.currency)
+		})
+	}
+}
+
+func TestFormatNumberByParts_USD(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(12345.67, NumberFormatOptions{
+		Currency: "USD",
+	})
+	require.NoError(t, err)
+
+	assert.Equal(t, "$", result.Currency)
+	assert.Equal(t, "12,345", result.Integer)
+	assert.Equal(t, ".", result.Decimal)
+	assert.True(t, result.IsPrefixSymbol)
+	assert.Equal(t, PartCurrency, result.RawParts[0].Type)
+	assert.Equal(t, "$", result.RawParts[0].Value)
+}
+
+func TestFormatNumberByParts_EUR(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(12345.67, NumberFormatOptions{
+		Currency: "EUR",
+	})
+	require.NoError(t, err)
+
+	assert.Equal(t, "€", result.Currency)
+	assert.Equal(t, "12,345", result.Integer)
+	assert.Equal(t, ".", result.Decimal)
+	assert.True(t, result.IsPrefixSymbol)
+	assert.Equal(t, PartCurrency, result.RawParts[0].Type)
+	assert.Equal(t, "€", result.RawParts[0].Value)
+}
+
+func TestFormatNumberByParts_Negative_Prefix(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(-500.0, NumberFormatOptions{
+		Currency: "USD",
+	})
+	require.NoError(t, err)
+
+	assert.Equal(t, "-", result.MinusSign)
+	assert.True(t, result.IsPrefixSymbol)
+	assert.Equal(t, PartMinusSign, result.RawParts[0].Type)
+	assert.Equal(t, PartCurrency, result.RawParts[1].Type)
+}
+
+func TestFormatNumberByParts_Negative_Prefix_EUR(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(-1234567.0, NumberFormatOptions{
+		Currency: "EUR",
+	})
+	require.NoError(t, err)
+
+	assert.Equal(t, "-", result.MinusSign)
+	assert.True(t, result.IsPrefixSymbol)
+	assert.Equal(t, PartMinusSign, result.RawParts[0].Type)
+	assert.Equal(t, PartCurrency, result.RawParts[1].Type)
+}
+
+func TestFormatNumberByParts_NoCurrency(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(123.45, NumberFormatOptions{})
+	require.NoError(t, err)
+
+	assert.Equal(t, "123", result.Integer)
+	assert.Equal(t, ".", result.Decimal)
+	assert.Empty(t, result.Currency)
+	assert.True(t, result.IsPrefixSymbol)
+}
+
+func TestFormatNumberByParts_JPY(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(12346.0, NumberFormatOptions{
+		Currency: "JPY",
+	})
+	require.NoError(t, err)
+
+	assert.Empty(t, result.Decimal)
+	assert.NotEmpty(t, result.Integer)
+}
+
+func TestFormatNumberByParts_InvalidAmount(t *testing.T) {
+	_, err := cachedCurrencyData.FormatNumberByParts("not-a-number", NumberFormatOptions{
+		Currency: "USD",
+	})
+	assert.Error(t, err)
+}
+
+func TestFormatNumberByParts_UnknownCurrency(t *testing.T) {
+	result, err := cachedCurrencyData.FormatNumberByParts(12345.67, NumberFormatOptions{
+		Currency: "XYZ",
+	})
+	require.NoError(t, err)
+
+	assert.Equal(t, "XYZ", result.Currency)
+	assert.True(t, result.IsPrefixSymbol)
+
+	require.GreaterOrEqual(t, len(result.RawParts), 2)
+	assert.Equal(t, PartCurrency, result.RawParts[0].Type)
+	assert.Equal(t, PartLiteral, result.RawParts[1].Type)
 }
